+++ core/serializers.py
@@ -1,0 +1,4 @@
+from rest_framework import serializers
+
+class DelaySerializer(serializers.Serializer):
+    time = serializers.IntegerField(default=10)