+++ rpirest/celery.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 import os
-from celery import Celery
+from celery import Celery, shared_task
+import time
 
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'rpirest.settings')
@@ -20,3 +21,7 @@ app.autodiscover_tasks()
 @app.task(bind=True)
 def debug_task(self):
     print('Request: {0!r}'.format(self.request))
+
+@shared_task
+def sleep(sleep_time):
+    time.sleep(sleep_time)